[WIP] Huge numbers (#472)
* Create test_huge_number.py


* style: format code with Autopep8, Black, Go fmt, Gofumpt, isort and Yapf

This commit fixes the style issues introduced in b91775b according to the output
from Autopep8, Black, Go fmt, Gofumpt, isort and Yapf.

Details: https://github.com/ianfhunter/GNOLL/pull/472

* Update test_huge_number.py


* style: format code with Autopep8, Black, Go fmt, Gofumpt, isort and Yapf

This commit fixes the style issues introduced in bf8f6a3 according to the output
from Autopep8, Black, Go fmt, Gofumpt, isort and Yapf.

Details: https://github.com/ianfhunter/GNOLL/pull/472

* Update test_huge_number.py


* Update safe_functions.c


* Update safe_functions.h


* Update dice.lex


* Update vec.h


* Update roll_parameters.h


* Update dice.yacc


* Update vec.h


* Update vector_functions.h


* Update vector_functions.c


* Update string_functions.h


* Update vector_functions.h


* Update string_functions.c


* Update vector_functions.c


* Update safe_functions.h


* Update safe_functions.c


* Update mocking.h


* Update mocking.c


* Update array_functions.h


* Update array_functions.c


* Update randomness.c


* Update randomness.h


* Update dice_frontend.c


* Update dice_core.h


* Update dice_core.c


* Update vec.h


* Update dice_frontend.c


* Update vec.h


* Update vec.h


* Update dice_frontend.c


* Update dice_frontend.c


* Update dice.yacc


* Update dice.lex


* Update shared_header.h


* Update vector_functions.h


* Update vector_functions.c


* Update safe_functions.c


* Update array_functions.h


* Update array_functions.c


* Update dice_frontend.h


* Update dice_frontend.c


* Update roll_parameters.h


* Update vector_functions.c


* Update vector_functions.c


* Update vector_functions.c


* Update dice.yacc


* Update macros.c


* Update conditionals.c


* Update conditionals.h


* Update vector_functions.c


* Update vector_functions.h


* Update dice.yacc


* Update conditionals.c


* Update vector_functions.c


* Update conditionals.c


* Update conditionals.h


* replace pcg-basic with pcg-c

* Update Makefile


* Update test_OS.yml


* Update test_core.yml


* Update test_language_bindings.yml


* Update test_packages.yml


* Update test_OS.yml


* Update randomness.c


* Update Makefile


* Update dice.yacc


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update Makefile


* Update randomness.c


* Update Makefile


* Update safe_functions.h


* Update dice.yacc


* Update safe_functions.c


* Update safe_functions.h


* Update safe_functions.c


* Update dice.yacc


* Update target.mk


* Update target.mk


* Update Makefile


* Update target.mk


* Update target.mk


* Update test_language_bindings.yml


* Update randomness.c


* Update safe_functions.c


* Update safe_functions.h


* Update lint_codeQL.yml


* Update target.mk


* Update target.mk


* Update randomness.c


* Update target.mk


* Update test_OS.yml


* Update target.mk


* Update test_huge_number.py


* Update target.mk


* Update target.mk


* Update target.mk


* Update target.mk


* Update target.mk


* Update test_troika.py


* Update target.mk


* remove union to another PR

* Update dice.yacc


* Update conditionals.c


* Update dice.yacc


* Update dice.yacc


* Update dice.yacc


* Update dice_core.c


* Update dice.yacc


* Update __init__.py


* style: format code with Autopep8, Black, Go fmt, Gofumpt, isort and Yapf

This commit fixes the style issues introduced in 7365bee according to the output
from Autopep8, Black, Go fmt, Gofumpt, isort and Yapf.

Details: https://github.com/ianfhunter/GNOLL/pull/472

* Update __init__.py


* style: format code with Autopep8, Black, Go fmt, Gofumpt, isort and Yapf

This commit fixes the style issues introduced in 346fe6d according to the output
from Autopep8, Black, Go fmt, Gofumpt, isort and Yapf.

Details: https://github.com/ianfhunter/GNOLL/pull/472

* Update test_huge_number.py


* Update test_huge_number.py


* Update dice.yacc


* Update src/grammar/rolls/dice_core.c


* Update src/grammar/rolls/dice_core.c


* Update test_huge_number.py


* style: format code with Autopep8, Black, Go fmt, Gofumpt, isort and Yapf

This commit fixes the style issues introduced in 85e22cf according to the output
from Autopep8, Black, Go fmt, Gofumpt, isort and Yapf.

Details: https://github.com/ianfhunter/GNOLL/pull/472

* Update test_huge_number.py


* Update __init__.py


* style: format code with Autopep8, Black, Go fmt, Gofumpt, isort and Yapf

This commit fixes the style issues introduced in 598b67c according to the output
from Autopep8, Black, Go fmt, Gofumpt, isort and Yapf.

Details: https://github.com/ianfhunter/GNOLL/pull/472

* Update randomness.c


* Update randomness.c


* Update randomness.h


* Update randomness.c


* Update randomness.c


* Update randomness.h


* Update randomness.c


* Update randomness.h


* Update test_huge_number.py


---------

+++ src/grammar/constructs/roll_parameters.h
@@ -4,11 +4,11 @@
 #include "constructs/dice_enums.h"
 
 typedef struct roll_params {
-  unsigned int number_of_dice;
-  unsigned int die_sides;
+  unsigned long long number_of_dice;
+  unsigned long long die_sides;
   EXPLOSION_TYPE explode;
   char** symbol_pool;
-  int start_value;
+  long long start_value;
   DIE_TYPE dtype;
 } roll_params;
