+++ src/grammar/util/array_functions.c
@@ -1,12 +1,12 @@
 
 #include "util/array_functions.h"
 
-int collapse(int * arr, unsigned int len){
+long long collapse(long long * arr, unsigned long long len){
     return sum(arr, len);
 }
 
-int sum(int * arr, unsigned int len){
-    int result = 0;
-    for(unsigned int i = 0; i != len; i++) result += arr[i];
+long long sum(long long * arr, unsigned long long len){
+    long long result = 0;
+    for(unsigned long long i = 0; i != len; i++) result += arr[i];
     return result;
 }
